+++ crane/cli.go
@@ -127,6 +127,10 @@ var (
 		"start",
 		"Start the containers.",
 	)
+	startAttachFlag = startCommand.Flag(
+		"attach",
+		"Attach to container",
+	).Short('a').Bool()
 	startTargetArg = startCommand.Arg("target", "Target of command").String()
 
 	stopCommand = app.Command(
@@ -356,7 +360,7 @@ func runCli() {
 
 	case startCommand.FullCommand():
 		commandAction(*startTargetArg, func(uow *UnitOfWork) {
-			uow.Start()
+			uow.Start(*startAttachFlag)
 		}, true)
 
 	case stopCommand.FullCommand():